Function ex00::ft_strcpy

pub unsafe fn ft_strcpy(dst: *mut u8, src: *const u8)

Copies proper C-style string (null-terminated) at src into dst

Safety

src must be: Valid readable null-terminated string. Accessible and not be deallocated during operation.

dst must be: Valid writable memory with sufficient space (dst has at least strlen(src) + 1 allocated bytes).

src and dst must not overlap

This function does not perform any bounds checking, and the caller must ensure that both pointers are valid and point to accessible memory.